Delivered fix for capturing screenshots on failures (move onto the stage where drivers are still actual
